requirements.txt
# A2A Multi-Agent Orchestrator System Requirements

# Core A2A SDK
a2a-python>=0.1.0

# Web Frameworks
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# HTTP Client
httpx>=0.25.0

# Environment & Configuration
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# Authentication
pyjwt>=2.8.0

# Type Extensions
typing-extensions>=4.8.0